+++ tests/test_visitors/test_ast/test_decorators/test_new_style_decorators.py
@@ -1,5 +1,6 @@
 import pytest
 
+from wemake_python_styleguide.compat.constants import PY312
 from wemake_python_styleguide.violations.best_practices import (
     NewStyledDecoratorViolation,
 )
@@ -18,6 +19,39 @@ class Some:
     def some(self): ...
 """
 
+invalid_decorators = [
+    'some[1]',
+    'some.attr[0]',
+    'some[0].attr',
+    'call()[1].attr',
+    'some + other',
+    'really @ strange[0]',
+]
+
+valid_decorators = [
+    'some',
+    'some()',
+    'some(index[1])',
+    'some.attr',
+    'some.attr(1 + 1)',
+]
+
+invalid_decorators3_12 = [
+    'some + other',
+    'some[1] + other[1]',
+    'some.attr + other.attr',
+    'some[0].attr + other.attr[0]',
+    'really @ strange[0]',
+]
+
+valid_decorators3_12 = [
+    *valid_decorators,
+    'some[my_type]',
+    'some[my_type](index)',
+    'some.attr[my_type]()',
+    'some.attr(1)[my_type]',
+]
+
 
 @pytest.mark.parametrize(
     'code',
@@ -28,14 +62,7 @@ class Some:
 )
 @pytest.mark.parametrize(
     'decorator',
-    [
-        'some[1]',
-        'some.attr[0]',
-        'some[0].attr',
-        'call()[1].attr',
-        'some + other',
-        'really @ strange[0]',
-    ],
+    invalid_decorators3_12 if PY312 else invalid_decorators,
 )
 def test_invalid_decorators(
     assert_errors,
@@ -63,13 +90,7 @@ def test_invalid_decorators(
 )
 @pytest.mark.parametrize(
     'decorator',
-    [
-        'some',
-        'some()',
-        'some(index[1])',
-        'some.attr',
-        'some.attr(1 + 1)',
-    ],
+    valid_decorators3_12 if PY312 else valid_decorators,
 )
 def test_valid_decorators(
     assert_errors,