+++ tests/test_visitors/test_ast/test_subscripts/test_redundant_subscripts.py
@@ -12,7 +12,6 @@ usage_template = 'constant[{0}]'
     'expression',
     [
         '0:7',
-        '0:7:1',
         'None:7',
         '3:None',
         '3:None:2',
@@ -20,7 +19,7 @@ usage_template = 'constant[{0}]'
         '3:7:1',
     ],
 )
-def test_redundant_subscript(
+def test_one_redundant_subscript(
     assert_errors,
     parse_ast_tree,
     expression,
@@ -33,6 +32,70 @@ def test_redundant_subscript(
     visitor.run()
 
     assert_errors(visitor, [RedundantSubscriptViolation])
+
+
+@pytest.mark.parametrize(
+    'expression',
+    [
+        '0:7:1',
+        '0:7:None',
+        '0:None',
+        'None:7:None',
+        'None:None',
+        '3:None:1',
+        ':None:None',
+    ],
+)
+def test_two_redundant_subscript(
+    assert_errors,
+    parse_ast_tree,
+    expression,
+    default_options,
+):
+    """Testing that redundant subscripts are forbidden."""
+    tree = parse_ast_tree(usage_template.format(expression))
+
+    visitor = SubscriptVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(
+        visitor,
+        [
+            RedundantSubscriptViolation,
+            RedundantSubscriptViolation,
+        ],
+    )
+
+
+@pytest.mark.parametrize(
+    'expression',
+    [
+        '0:None:1',
+        'None:None:1',
+        'None:None:None',
+        '0:None:None',
+    ],
+)
+def test_three_redundant_subscript(
+    assert_errors,
+    parse_ast_tree,
+    expression,
+    default_options,
+):
+    """Testing that redundant subscripts are forbidden."""
+    tree = parse_ast_tree(usage_template.format(expression))
+
+    visitor = SubscriptVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(
+        visitor,
+        [
+            RedundantSubscriptViolation,
+            RedundantSubscriptViolation,
+            RedundantSubscriptViolation,
+        ],
+    )
 
 
 @pytest.mark.parametrize(