+++ wemake_python_styleguide/visitors/ast/subscripts.py
@@ -57,26 +57,30 @@ class SubscriptVisitor(base.BaseNodeVisitor):
         if not isinstance(node.slice, ast.Slice):
             return
 
+        indexes: list[ast.expr | None] = []
         lower_ok = node.slice.lower is None or (
             not self._is_zero(node.slice.lower)
             and not self._is_none(node.slice.lower)
         )
-
         upper_ok = node.slice.upper is None or not self._is_none(
             node.slice.upper,
         )
-
         step_ok = node.slice.step is None or (
             not self._is_one(node.slice.step)
             and not self._is_none(node.slice.step)
         )
 
-        if not (lower_ok and upper_ok and step_ok):
-            self.add_violation(
-                consistency.RedundantSubscriptViolation(
-                    node,
-                ),
-            )
+        if not lower_ok:
+            indexes.append(node.slice.lower)
+
+        if not upper_ok:
+            indexes.append(node.slice.upper)
+
+        if not step_ok:
+            indexes.append(node.slice.step)
+
+        for index in indexes:
+            self.add_violation(consistency.RedundantSubscriptViolation(index))
 
     def _check_slice_assignment(self, node: ast.Subscript) -> None:
         if not isinstance(node.ctx, ast.Store):