+++ wemake_python_styleguide/options/config.py
@@ -155,6 +155,9 @@ You can also show all options that ``flake8`` supports by running:
 - ``max-match-cases`` - maximum number of cases in a match block of code
     defaults to
     :str:`wemake_python_styleguide.options.defaults.MAX_MATCH_CASES`
+- ``max-lines-in-finally`` - maximum lines in finally block of code
+    defaults to
+    :str:`wemake_python_styleguide.options.default.MAX_LINES_IN_FINALLY
 
 .. rubric:: Formatter options
 
@@ -430,6 +433,11 @@ class Configuration:
             '--max-match-cases',
             defaults.MAX_MATCH_CASES,
             'Maximum number of match cases in a single match.',
+        ),
+        _Option(
+            '--max-lines-in-finally',
+            defaults.MAX_LINES_IN_FINALLY,
+            'Maximum lines of expressions in a finally block.'
         ),
         # Formatter:
         _Option(