+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+commit b4ccf16117cf441468fdcadeab671cae6a040a1f
+Author: Alexeev Bronislav <alexeev.dev@mail.ru>
+Date:   Tue Jun 17 03:13:06 2025 +0700
+
+    add Untrivial Logic in Finally block checker
+
 commit c2721c0f6aebbc30940219e8ff38df9ccaa9394f
 Author: Alexeev Bronislav <alexeev.dev@mail.ru>
 Date:   Tue Jun 17 02:13:56 2025 +0700